+++ StarkInfra/StarkInfra/IssuingCard/IssuingCard.cs
@@ -35,6 +35,7 @@ namespace StarkInfra
     ///     <item>HolderID [string]: card holder unique id. ex: "5656565656565656"</item>
     ///     <item>Type [string]: card type. ex: "virtual"</item>
     ///     <item>Status [string]: current IssuingCard status. ex: "active", "blocked", "canceled" or "expired"</item>
+    ///     <item>IsPinDefined [bool]: Whether the card has a PIN defined. Returned only when "expand=isPinDefined" is informed in the request</item>
     ///     <item>Number [string]: [EXPANDABLE] masked card number. Expand to unmask the value. ex: "123".</item>
     ///     <item>SecurityCode [string]: [EXPANDABLE] masked card verification value (cvv). Expand to unmask the value. ex: "123".</item>
     ///     <item>Expiration [DateTime]: [EXPANDABLE] masked card expiration DateTime. Expand to unmask the value. ex: DateTime(2020, 3, 10, 10, 30, 0, 0).</item>
@@ -49,7 +50,7 @@ namespace StarkInfra
         public string HolderExternalID { get; }
         public string DisplayName { get; }
         public List<IssuingRule> Rules { get; }
-        public string BinID { get; }
+        public string ProductID { get; }
         public List<string> Tags { get; }
         public string StreetLine1 { get; }
         public string StreetLine2 { get; }
@@ -60,6 +61,7 @@ namespace StarkInfra
         public string HolderID {  get; }
         public string Type { get; }
         public string Status { get; }
+        public bool? IsPinDefined { get; }
         public string Number { get; }
         public string SecurityCode { get; }
         public DateTime? Expiration { get; }
@@ -101,6 +103,7 @@ namespace StarkInfra
         ///     <item>holderID [string]: card holder unique id. ex: "5656565656565656"</item>
         ///     <item>type [string]: card type. ex: "virtual"</item>
         ///     <item>status [string]: current IssuingCard status. ex: “active”, “blocked”, “canceled” or “expired"</item>
+        ///     <item>isPinDefined [bool]: Whether the card has a PIN defined. Returned only when "expand=isPinDefined" is informed in the request</item>
         ///     <item>number [string]: [EXPANDABLE] masked card number. Expand to unmask the value. ex: "123".</item>
         ///     <item>securityCode [string]: [EXPANDABLE] masked card verification value (cvv). Expand to unmask the value. ex: "123".</item>
         ///     <item>expiration [DateTime]: [EXPANDABLE] masked card expiration DateTime. Expand to unmask the value. ex: DateTime(2020, 3, 10, 10, 30, 0, 0).</item>
@@ -110,8 +113,8 @@ namespace StarkInfra
         /// </summary>
         public IssuingCard(string holderName, string holderTaxID, string holderExternalID, string displayName = null, List<IssuingRule> rules = null, 
             string productID = null, List<string> tags = null, string streetLine1 = null, string streetLine2 = null, string district = null, string city = null, 
-            string stateCode = null, string zipCode = null, string id = null, string holderID = null, string type = null, string status = null, 
-            string number = null, string securityCode = null, DateTime? expiration = null, DateTime? updated = null, DateTime? created = null
+            string stateCode = null, string zipCode = null, string id = null, string holderID = null, string type = null, string status = null,
+            bool? isPinDefined = null, string number = null, string securityCode = null, DateTime? expiration = null, DateTime? updated = null, DateTime? created = null
         ) : base(id)
         {
             HolderID = holderID;
@@ -121,8 +124,9 @@ namespace StarkInfra
             Type = type;
             DisplayName = displayName;
             Status = status;
+            IsPinDefined = isPinDefined;
             Rules = rules;
-            BinID = productID;
+            ProductID = productID;
             StreetLine1 = streetLine1;
             StreetLine2 = streetLine2;
             District = district;
@@ -337,6 +341,7 @@ namespace StarkInfra
                     { "ids", ids },
                     { "after", after },
                     { "before", before },
+                    { "status", status },
                     { "types", types },
                     { "holderIds", holderIds },
                     { "tags", tags },
@@ -431,7 +436,7 @@ namespace StarkInfra
             string holderExternalID = json.holderExternalId;
             string displayName = json.displayName;
             List<IssuingRule> rules = IssuingRule.ParseRules(json.rules);
-            string productID = json.binId;
+            string productID = json.productId;
             List<string> tags = json.tags?.ToObject<List<string>>();
             string streetLine1 = json.streetLine1;
             string streetLine2 = json.streetLine2;
@@ -443,6 +448,7 @@ namespace StarkInfra
             string holderID = json.holderId;
             string type = json.type;
             string status = json.status;
+            bool? isPinDefined = json.isPinDefined;
             string number = json.number;
             string securityCode = json.securityCode;
             string expirationString = json.expiration;
@@ -457,8 +463,8 @@ namespace StarkInfra
                 holderName: holderName, holderTaxID: holderTaxID, holderExternalID: holderExternalID, 
                 displayName: displayName, rules: rules, productID: productID, tags: tags, 
                 streetLine1: streetLine1, streetLine2: streetLine2, district: district, 
-                city: city, stateCode: stateCode, zipCode: zipCode, id: id, holderID: holderID, 
-                type: type, status: status, number: number, securityCode: securityCode, 
+                city: city, stateCode: stateCode, zipCode: zipCode, id: id, holderID: holderID,
+                type: type, status: status, isPinDefined: isPinDefined, number: number, securityCode: securityCode,
                 expiration: expiration, updated: updated, created: created
             );
         }