+++ CHANGELOG.md
@@ -13,6 +13,15 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- productId attribute to IssuingCard resource
+- type attribute to IssuingDesign resource
+- productId, merchantCategoryType, installmentCount, description and holderId attributes to IssuingPurchase resource
+- IssuingBillingInvoice resource
+- IssuingBillingTransaction resource
+- IssuingToken resource
+- IssuingTokenRequest resource
+- IssuingTokenDesign resource
 
 ## [0.13.0] - 2026-06-29
 ### Added